refactor: 도메인 모델 패키지 생성
