+++ app/src/main/java/nextstep/signup/model/Password.kt
@@ -1,4 +1,4 @@
-package nextstep.signup
+package nextstep.signup.model
 
 data class Password(val password: String) {
     fun isValid() = !isBlank() && (!isValidLength() || !hasEnglishAndNumber())
